♻️ UpperDashbordのpropsをトップページから持ってくるようにした (#49)
